+++ packages/tree-view/src/presenters/SubTreeViewCombined.js
@@ -224,6 +224,7 @@ const SubTreeViewCombined = props => {
       children,
       // eslint-disable-next-line no-shadow
       collapsed,
+      defaultSelected,
       density,
       getActiveTreeItemId,
       getActiveTreeItemIndex,
@@ -233,7 +234,9 @@ const SubTreeViewCombined = props => {
       guidelines,
       id,
       indicator,
+      isControlled,
       level,
+      selected,
       setActiveTreeItemId,
       setActiveTreeItemIndex,
       setKeyboardOpenId,
@@ -245,6 +248,7 @@ const SubTreeViewCombined = props => {
     const clonedChildren = Array.isArray(children)
       ? children.map(child =>
           React.cloneElement(child, {
+            defaultSelected,
             getActiveTreeItemId,
             getActiveTreeItemIndex,
             getCurrentItemClicked,
@@ -252,14 +256,17 @@ const SubTreeViewCombined = props => {
             getTreeItemArray,
             guidelines,
             indicator,
+            isControlled,
             keyboardOpenId: getKeyboardOpenId(),
             level: level + 1,
+            selected,
             setActiveTreeItemId,
             setActiveTreeItemIndex,
             setKeyboardOpenId
           })
         )
       : React.cloneElement(children, {
+          defaultSelected,
           getActiveTreeItemId,
           getActiveTreeItemIndex,
           getCurrentItemClicked,
@@ -267,8 +274,10 @@ const SubTreeViewCombined = props => {
           getTreeItemArray,
           guidelines,
           indicator,
+          isControlled,
           keyboardOpenId: getKeyboardOpenId(),
           level: level + 1,
+          selected,
           setActiveTreeItemId,
           setActiveTreeItemIndex,
           setKeyboardOpenId