+++ packages/tree-view/src/presenters/fileview/TreeObjectSubTreeItem.js
@@ -15,11 +15,14 @@ const TreeObjectSubTreeItem = props => {
       id,
       meta: { className, icon, label },
       payload: {
+        defaultSelected,
         getActiveTreeItemId,
         getActiveTreeItemIndex,
         getCurrentItemClicked,
         guidelines,
-        indicator
+        indicator,
+        isControlled,
+        selected
       },
       onClick: userOnClick,
       ...otherTreeItemProps
@@ -30,6 +33,12 @@ const TreeObjectSubTreeItem = props => {
     ...otherProps
   } = props;
   const { onMouseEnter, onMouseLeave } = otherProps;
+  const getActiveId = () => {
+    if (isControlled()) {
+      return selected;
+    }
+    return getCurrentItemClicked() || defaultSelected;
+  };
   const styleTreeItem = {
     children,
     id,
@@ -41,7 +50,7 @@ const TreeObjectSubTreeItem = props => {
     getActiveTreeItemId,
     getActiveTreeItemIndex,
     guidelines,
-    selected: getCurrentItemClicked() === id
+    selected: getActiveId() === id
   };
   const { handleClick, handleKeyDown } = createButtonEventHandlers(onClick);
   const higTreeItemContentWrapperClassName = createCustomClassNames(